% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{autoAnnoResult}
\alias{autoAnnoResult}
\title{autoAnnoResult
Integrate the prediction results of the algorithm in the autoAnnoTools function.}
\usage{
autoAnnoResult(anno)
}
\arguments{
\item{anno}{The predicted results of the algorithm in the autoAnnoTools function.}
}
\value{
Seurat object with consolidated results.
}
\description{
autoAnnoResult
Integrate the prediction results of the algorithm in the autoAnnoTools function.
}
