% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{autoAnnoTools}
\alias{autoAnnoTools}
\title{autoAnnoTools}
\usage{
autoAnnoTools(
  obj.seu,
  ref.obj = NULL,
  ref.ctype = NULL,
  marker.lst = NULL,
  method = c("SingleR", "Seurat", "sciBet", "scmap", "CHETAH", "scSorter", "sc.type",
    "cellID", "scCATCH", "SCINA"),
  select.marker = c("Seurat"),
  top.k = 30,
  strategy = c("reference-based", "marker-based"),
  ...
)
}
\arguments{
\item{obj.seu}{Seurat object, which needs to be annotated.}

\item{ref.obj}{Seurat object, only when used with reference-based tools. Default: NULL.}

\item{marker.lst}{A list contained maker genes for each cell type for marker based tool.
Notably, ref.obj and marker.lst are antagonistic, and must be provide one. Default: NULL.}

\item{method}{A vector of automated annotation tools.}

\item{select.marker}{Specify method for inferring markers for each subset. Default: Seurat.}

\item{top.k}{Top k expressed genes of each subset remained. Default: NULL.}
}
\value{
Annotated Seurat object, predicted results are embedded into meta.data slot.
}
\description{
Automated single-cell annotation using publicly available tools.
}
