% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoTools.R
\name{cellIDAnno}
\alias{cellIDAnno}
\title{cellIDAnno
Automated single-cell annotation using CellID (Ianevski et al., 2022).}
\usage{
cellIDAnno(
  obj.seu,
  marker.lst = NULL,
  gset.len = 5,
  sig.cut = 2,
  species = c("Hs", "Mm"),
  tissue.type = NULL,
  ...
)
}
\arguments{
\item{obj.seu}{Query Seurat object, which needs to be annotated.}

\item{marker.lst}{A list contained maker genes for each cell type.}

\item{gset.len}{Cell types with a number of markers less than k need to be removed. Default: 5.}

\item{sig.cut}{Optimal cutoff value to determine significant prediction or not. Default: 2.}

\item{species}{Specify species of the query Seurat object data when using markes from PanglaoDB database. Default: Hs.}

\item{tissue.type}{Specify the tissue type when using markes from PanglaoDB database. Default: NULL.}

\item{...}{More arguments can be assessed using the sctype_score function in the sc.type package.}
}
\value{
Annotated Seurat object, predicted results are embedded into meta.data slot.
}
\description{
cellIDAnno
Automated single-cell annotation using CellID (Ianevski et al., 2022).
}
\examples{
obj.seu <- system.file("data/test", "test.obj.rds", package = "Biotools") \%>\% readRDS(.)
ref.obj <- system.file("data/test", "ref.obj.rds", package = "Biotools") \%>\% readRDS(.)
marker.lst <- findMarkerToolsForSc(ref.obj, to.list = TRUE, top.k = 30)
obj.seu <- cellIDAnno(obj.seu, marker.lst)
}
