% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoTools.R
\name{chetahAnno}
\alias{chetahAnno}
\title{chetahAnno}
\usage{
chetahAnno(obj.seu, ref.obj, ref.ctype = NULL, slot.data = "data", ...)
}
\arguments{
\item{obj.seu}{Query Seurat object, which needs to be annotated.}

\item{ref.obj}{Reference Seurat object.}

\item{ref.ctype}{Please specify information about the cell type in the reference object, otherwise: defaults to NULL.
i.e. Idents are extracted set to the cell type.}

\item{slot.data}{Extract expression in Seurat slot. Default: "data".}

\item{...}{More arguments can be assessed using the CHETAHclassifier function in the CHETAH package.}
}
\value{
Annotated Seurat object, predicted results are embedded into meta.data slot.
}
\description{
chetahAnno
}
\examples{
obj.seu <- system.file('data/test', 'test.obj.rds', package = 'Biotools') \%>\% readRDS(.)
ref.obj <- system.file('data/test', 'ref.obj.rds', package = 'Biotools') \%>\% readRDS(.)
obj.seu <- chetahAnno(obj.seu, ref.obj)
}
