% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarkerTools.R
\name{collapseLisToFrame}
\alias{collapseLisToFrame}
\title{collapseLisToFrame
Collapse the list to data.frame, which contains two coloumns, one is list names, and another is correponding values.}
\usage{
collapseLisToFrame(marker.lst)
}
\arguments{
\item{marker.lst}{A list with names.}
}
\value{
A collapsed data.frame.
}
\description{
collapseLisToFrame
Collapse the list to data.frame, which contains two coloumns, one is list names, and another is correponding values.
}
\examples{
my.list <- list(a = c(1, 2, 3), b = c("foo", "bar"), c = 4:6)
collapseLisToFrame(my.list)
}
