% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarkerTools.R
\name{commonClustering}
\alias{commonClustering}
\title{commonClustering
A function used for data preprocessing.}
\usage{
commonClustering(obj, res = 0.8)
}
\arguments{
\item{obj}{Seurat object.}

\item{res}{Default: 0.8.}
}
\value{
Seurat object.
}
\description{
commonClustering
A function used for data preprocessing.
}
\examples{
obj.seu <- system.file("data/test", "test.obj.rds", package = "Biotools") \%>\% readRDS(.)
obj.seu <- commonClustering(obj.seu)
}
