% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarkerTools.R
\name{downSamplSeurat}
\alias{downSamplSeurat}
\title{downSamplSeurat
A subsampling function for Seurat type data.}
\usage{
downSamplSeurat(obj, seed = 123, percent = 0.5)
}
\arguments{
\item{obj}{Seurat object.}

\item{seed}{Number of seed. Default: 123.}

\item{percent}{Number of percent. Default: 0.5.}
}
\description{
downSamplSeurat
A subsampling function for Seurat type data.
}
