% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarkerTools.R
\name{findMarkersBySeurat}
\alias{findMarkersBySeurat}
\title{findMarkersBySeurat
Find markers for each subset using the wilcox.test function wrapped in the FindAllMarkers function from the Seurat package.}
\usage{
findMarkersBySeurat(obj.seu, to.list = TRUE, top.k = NULL, ...)
}
\arguments{
\item{obj.seu}{Seurat object.}

\item{to.list}{Convert the results to a list or not. If TRUE, the names of the list represent the subset, and values are marker genes.}

\item{top.k}{Top k expressed genes of each subset remained. Default: NULL.}

\item{...}{More arguments can be assessed using the FindAllMarkers function in the Seurat package.}
}
\value{
A list of makers for the subsets or default format results generated by Seurat package.
}
\description{
findMarkersBySeurat
Find markers for each subset using the wilcox.test function wrapped in the FindAllMarkers function from the Seurat package.
}
\examples{
obj.seu <- system.file("data/test", "test.obj.rds", package = "Biotools") \%>\% readRDS(.)
results <- findMarkersBySeurat(obj.seu, to.list = TRUE, logfc.threshold = 0.25, min.pct = 0.25)
}
