% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarkerTools.R
\name{findMarkerToolsForSc}
\alias{findMarkerToolsForSc}
\title{findMarkerToolsForSc
Find markers for subsets using publicly avaiable tools on the basis of single-cell data.}
\usage{
findMarkerToolsForSc(
  obj.seu,
  to.list = TRUE,
  method = c("Seurat"),
  top.k = NULL,
  n.workers = 4,
  ...
)
}
\arguments{
\item{obj.seu}{Seurat object.}

\item{to.list}{Convert the results to a list or not. If TRUE, the names of the list represent the subset, and values are marker genes.}

\item{method}{A vector of publicly avaiable tools. Default: Seurat.}

\item{top.k}{Top k expressed genes of each subset remained. Default: NULL.}

\item{n.workers}{Number of cores for parallel when selected tool supported. Default: 4.}

\item{...}{More arguments can be assessed from the corresponding tools.}
}
\value{
A list of makers for the subsets or default format results generated by the selected tool.
}
\description{
findMarkerToolsForSc
Find markers for subsets using publicly avaiable tools on the basis of single-cell data.
}
\examples{
obj.seu <- system.file('data/test', 'test.obj.rds', package = 'Biotools') \%>\% readRDS(.)
results <- findMarkerToolsForSc(obj.seu, to.list = TRUE, method = 'Seurat', logfc.threshold = 0.25, min.pct = 0.25)
}
