% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{listToolMethods}
\alias{listToolMethods}
\title{listToolMethods}
\usage{
listToolMethods()
}
\value{
A vector functions of tools used for single-cell annotation.
}
\description{
Select the direction function of the single cell annotation tool.
}
