% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarkerTools.R
\name{multipleProcess}
\alias{multipleProcess}
\title{multipleProcess
Open multiple workers for Seurat processing.}
\usage{
multipleProcess(n.workers = 10)
}
\arguments{
\item{n.workers}{Number of cores. Default: 10.}
}
\description{
multipleProcess
Open multiple workers for Seurat processing.
}
\examples{
multipleProcess(n.workers = 10)
}
