% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarkerTools.R
\name{println}
\alias{println}
\title{println
Print prompt message.}
\usage{
println(X, verbose = TRUE, status = c("INFO", "WARN", "ERROR"))

println(X, verbose = TRUE, status = c("INFO", "WARN", "ERROR"))
}
\arguments{
\item{verbose}{Bool value, print out messages or not, default: TRUE.}

\item{status}{(optional) Status of running messages or error. Default: INFO.}

\item{infos}{Message that need to be printed.}
}
\value{
NULL.

NULL.
}
\description{
println
Print prompt message.

println
Print prompt message.
}
\examples{
println('Begin analysis ->', verbose = TRUE, status = 'INFO', 'example')
println('Begin analysis ->', verbose = TRUE, status = 'INFO', 'example')
}
