% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoTools.R
\name{sccatchAnno}
\alias{sccatchAnno}
\title{sccatchAnno
Automated single-cell annotation using scCATCH (Shao et al., 2020).}
\usage{
sccatchAnno(
  obj.seu,
  marker.lst = NULL,
  qry.cluster = "idents",
  species = c("Human", "Mouse"),
  ...
)
}
\arguments{
\item{obj.seu}{Query Seurat object, which needs to be annotated.}

\item{marker.lst}{A list contained maker genes for each cell type.}

\item{qry.cluster}{Specify cluster information of query Seurat object in meta.data slot. Default: idents.}

\item{species}{Species of query Seurat object data which needs to be annotated. Default: Human.}

\item{...}{More arguments can be assessed using the findcelltype function in the scCATCH package.}
}
\value{
Annotated Seurat object, predicted results are embedded into meta.data slot.
}
\description{
sccatchAnno
Automated single-cell annotation using scCATCH (Shao et al., 2020).
}
\examples{
obj.seu <- system.file("data/test", "test.obj.rds", package = "Biotools") \%>\% readRDS(.)
ref.obj <- system.file("data/test", "ref.obj.rds", package = "Biotools") \%>\% readRDS(.)
marker.lst <- findMarkerToolsForSc(ref.obj, to.list = TRUE, top.k = 30)
obj.seu <- sccatchAnno(obj.seu, marker.lst)
obj.seu <- sccatchAnno(obj.seu, marker.lst = NULL)
}
