% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoTools.R
\name{scinaAnno}
\alias{scinaAnno}
\title{scinaAnno}
\usage{
scinaAnno(obj.seu, marker.lst, ...)
}
\arguments{
\item{obj.seu}{Query Seurat object, which needs to be annotated.}

\item{marker.lst}{A list contained maker genes for each cell type.}

\item{...}{More arguments can be assessed using the scSorter function in the SCINA package.}
}
\value{
Annotated Seurat object, predicted results are embedded into meta.data slot.
}
\description{
scinaAnno
}
\examples{
obj.seu <- system.file('data/test', 'test.obj.rds', package = 'Biotools') \%>\% readRDS(.)
ref.obj <- system.file('data/test', 'ref.obj.rds', package = 'Biotools') \%>\% readRDS(.)
marker.lst <- findMarkerToolsForSc(ref.obj, to.list = TRUE, top.k = 30)
obj.seu <- scinaAnno(obj.seu, marker.lst)
}
