% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoTools.R
\name{scmapAnno}
\alias{scmapAnno}
\title{scmapAnno}
\usage{
scmapAnno(
  obj.seu,
  ref.obj,
  ref.ctype = NULL,
  slot.data = c("data", "count"),
  ...
)
}
\arguments{
\item{obj.seu}{Query Seurat object, which needs to be annotated.}

\item{ref.obj}{Reference Seurat object.}

\item{ref.ctype}{Please specify information about the cell type in the reference object, otherwise: defaults to NULL.
i.e. Idents are extracted set to the cell type.}

\item{slot.data}{Extract expression in Seurat slot. Default: "data".}

\item{...}{More arguments can be assessed using the scmapCell function in the scmap package.}
}
\value{
Annotated Seurat object, predicted results are embedded into meta.data slot.
}
\description{
scmapAnno
}
\examples{
obj.seu <- system.file("data/test", "test.obj.rds", package = "Biotools") \%>\% readRDS(.)
ref.obj <- system.file("data/test", "ref.obj.rds", package = "Biotools") \%>\% readRDS(.)
obj.seu <- scmapAnno(obj.seu, ref.obj)
}
