% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoTools.R
\name{sctypeAnno}
\alias{sctypeAnno}
\title{sctypeAnno
Automated single-cell annotation using sc.type (Ianevski et al., 2022).}
\usage{
sctypeAnno(obj.seu, marker.lst = NULL, tissue.type = "Immune system", ...)
}
\arguments{
\item{obj.seu}{Query Seurat object, which needs to be annotated.}

\item{marker.lst}{A list contained maker genes for each cell type. Default: NULL.}

\item{tissue.type}{Please specify the tissue type when using the default marker database wrapped by the sc.type package. Default: Immune system.}

\item{...}{More arguments can be assessed using the sctype_score function in the sc.type package.}
}
\value{
Annotated Seurat object, predicted results are embedded into meta.data slot.
}
\description{
sctypeAnno
Automated single-cell annotation using sc.type (Ianevski et al., 2022).
}
\examples{
obj.seu <- system.file("data/test", "test.obj.rds", package = "Biotools") \%>\% readRDS(.)
ref.obj <- system.file("data/test", "ref.obj.rds", package = "Biotools") \%>\% readRDS(.)
marker.lst <- findMarkerToolsForSc(ref.obj, to.list = TRUE, top.k = 30)
obj.seu <- sctypeAnno(obj.seu, marker.lst)
}
