% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoTools.R
\name{seuratAnno}
\alias{seuratAnno}
\title{seuratAnno
Automated single-cell annotation using Seurat (Stuart et al., 2019).}
\usage{
seuratAnno(obj.seu, ref.obj, ref.ctype = NULL, npcs = 30, ...)
}
\arguments{
\item{obj.seu}{Query Seurat object, which needs to be annotated.}

\item{ref.obj}{Reference Seurat object.}

\item{ref.ctype}{Please specify information about the cell type in the reference object, otherwise: defaults to NULL.
i.e. Idents are extracted set to the cell type.}

\item{npcs}{Number of PCs used for prediction. Default: 30.}

\item{...}{More arguments can be assessed using the TransferData function in the Seurat package.}
}
\value{
Annotated Seurat object, predicted results are embedded into meta.data slot.
}
\description{
seuratAnno
Automated single-cell annotation using Seurat (Stuart et al., 2019).
}
\examples{
obj.seu <- system.file("data/test", "test.obj.rds", package = "Biotools") \%>\% readRDS(.)
ref.obj <- system.file("data/test", "ref.obj.rds", package = "Biotools") \%>\% readRDS(.)
obj.seu <- seuratAnno(obj.seu, ref.obj)
}
