% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoTools.R
\name{singleRAnno}
\alias{singleRAnno}
\title{singleRAnno
Automated single-cell annotation using SingleR (Aran et al., 2019).}
\usage{
singleRAnno(obj.seu, ref.obj, ref.ctype = NULL, ...)
}
\arguments{
\item{obj.seu}{Query Seurat object, which needs to be annotated.}

\item{ref.obj}{Reference Seurat object.}

\item{ref.ctype}{Please specify information about the cell type in the reference object, otherwise: defaults to NULL.
i.e. Idents are extracted set to the cell type.}

\item{...}{More arguments can be assessed using the SingleR function in the SingleR package.}
}
\value{
Annotated Seurat object, predicted results are embedded into meta.data slot.
}
\description{
singleRAnno
Automated single-cell annotation using SingleR (Aran et al., 2019).
}
\examples{
obj.seu <- system.file("data/test", "test.obj.rds", package = "Biotools") \%>\% readRDS(.)
ref.obj <- system.file("data/test", "ref.obj.rds", package = "Biotools") \%>\% readRDS(.)
obj.seu <- singleRAnno(obj.seu, ref.obj)
}
